requirements.txt
flask==3.0.3
google-genai==0.3.0
pandas==2.1.4
